Access a workspace

Provides access to a workspace.

Usage

access_workspace(login, ws_id = NULL, ws_label = NULL, verbose = FALSE)

# S4 method for class 'LoginStudio'
access_workspace(login, ws_id = NULL, ws_label = NULL, verbose = FALSE)

# S4 method for class 'LoginTestcenter'
access_workspace(login, ws_id = NULL, ws_label = NULL, verbose = FALSE)

Arguments

login

Login Login object generated by either login_studio() or login_testcenter().

ws_id

Numeric. Workspace ws_id (see URL).

ws_label

Character. Workspace ws_label.

verbose

Logical. If TRUE, additional information is printed. Defaults to FALSE.

Value

Workspace object.

Functions

access_workspace(LoginStudio): Provide access to a selected workspace after logging in

access_workspace(LoginTestcenter): Provide access to a selected workspace after logging in