Download codebook

This function downloads codebooks from the IQB Studio.

Usage

download_codebook(
  workspace,
  path,
  unit_keys = NULL,
  format = "docx",
  missingsProfile = NULL,
  general_instructions = TRUE,
  only_manual = TRUE,
  closed = TRUE,
  derived = TRUE,
  show_score = FALSE,
  code_label_to_upper = TRUE
)

# S4 method for class 'WorkspaceStudio'
download_codebook(
  workspace,
  path,
  unit_keys = NULL,
  format = "docx",
  missingsProfile = NULL,
  general_instructions = TRUE,
  only_manual = TRUE,
  closed = TRUE,
  derived = TRUE,
  show_score = FALSE,
  code_label_to_upper = TRUE
)

Arguments

workspace

WorkspaceStudio. Workspace information necessary to download codebook via the API.

path

Character. Path for the codebook file to be downloaded.

unit_keys

Character. Keys (short names) of the units in the workspace the codebook should be retrieved from. If set to NULL (default), the codebook will be generated for the all units.

format

Character. Either "docx" (default) or json.

missingsProfile

Missings profile. (Currently with no effect.)

general_instructions

Logical. Should the general coding instructions be printed? Defaults to TRUE.

only_manual

Logical. Should only items with manual coding be printed? Defaults to TRUE.

closed

Logical. Should items that could be automatically coded be printed? Defaults to TRUE.

derived

Logical. Should the derived variables be printed? Defaults to TRUE.

show_score

Logical. Should the score be printed? Defaults to FALSE.

code_label_to_upper

Logical. Should the code labels be printed in capital letters? Defaults to TRUE.

Functions

download_codebook(WorkspaceStudio): Upload a file in a defined workspace